requirements.txt
# FINAL WORKING requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
streamlit==1.28.0
python-dotenv==1.0.0
duckduckgo-search==3.9.3
PyPDF2==3.0.0
python-docx==1.1.0
pandas==2.1.4
plotly==5.17.0

langchain==0.1.10
langchain-community==0.0.25 
langchain-openai==0.0.5

tiktoken==0.5.2


# Fixed versions for CrewAI + Azure compatibility
openai==1.55.3
httpx==0.27.2

# Additional dependencies from requirements_fresh.txt
httpx-sse==0.4.0
pydantic==2.5.0
pydantic-core==2.14.1
